requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
faiss-cpu
langchain
langchain_google_genai
streamlit
requests
PyMuPDF==1.24.1
langgraph